third_party/blink/web_tests/external/wpt/css/cssom-view/smooth-scroll-nonstop-expected.txt

This is a testharness.js-based test.
[FAIL] noop scrollIntoView doesn't interrupt ongoing smooth scroll.
  assert_equals: scroller reached the target offset expected 400 but got 0
Harness: the test ran to completion.
